requirements.txt
fastapi
black
uvicorn
httpx
ipdb
ruff
pan-os-python
icmplib
rich
napalm
h2
isort
